src/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.5.0

# Azure AI and authentication
openai
azure-ai-projects==1.0.0
azure-identity==1.15.0

# Additional dependencies that may be needed
python-multipart==0.0.6

python-dotenv
